# #450 Improve Marquee tool (Performance & Full element marquee behavior)