--- a/examples/workflow-standalone/src/di.config.ts
+++ b/examples/workflow-standalone/src/di.config.ts
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { createWorkflowDiagramContainer } from '@eclipse-glsp-examples/workflow-glsp/lib';
-import { GLSPDiagramServer } from '@eclipse-glsp/client';
+import { GLSPDiagramServer, GLSP_TYPES } from '@eclipse-glsp/client';
 import { Container } from 'inversify';
 import { ConsoleLogger, LogLevel, TYPES } from 'sprotty';
 import '../css/diagram.css';
@@ -25,5 +25,6 @@ export default function createContainer(): Container {
     container.bind(TYPES.ModelSource).toService(GLSPDiagramServer);
     container.rebind(TYPES.ILogger).to(ConsoleLogger).inSingletonScope();
     container.rebind(TYPES.LogLevel).toConstantValue(LogLevel.warn);
+    container.bind(GLSP_TYPES.IMarqueeBehavior).toConstantValue({ entireEdge: true, entireElement: true });
     return container;
 }