closes https://github.com/eclipse-glsp/glsp/issues/450

- Refactor code
  - Move calculation code to new class for easy testing
  - Return direct expression instead in if else
- Improve performance
  - Cache possible elements at start, instead of collect them on each mouse move event
  - Calculate absolute positions only once
- Add possibility to change the select behavior to marque full element
- Fix run test launch configs
- Add tests for calculation code of marquee tool

New marquee behavior:
![glsp-mark-entire-element](https://user-images.githubusercontent.com/42733123/140515253-765064ce-08a5-4582-9ea1-67a6171a10d1.gif)

I did again some performance measures in the MarqueeMouseListener class (same as in the Issue discribed):
* touch behavior: avg. 0.22880000030994416 ms
* touch behavior with 6x cpu slowdown: avg. 0.649664429910231 ms
* entire behavior: avg. 0.1789156626236726 ms
* entire behavior with 6x cpu slowdown: avg. 0.5752136751245229 ms

So it seems that the refactoring did its work 😀

Now the behavior is tested and can be changed over DI, e.g by:
`container.bind(GLSP_TYPES.IMarqueeBehavior).toConstantValue({ entireEdge: true, entireElement: true });`

Please add feedback if you want to be something changed or have an other improvement 😉