@planger 2021-11-08
Thank you very much @ivy-lli! This looks great!
I'll ask @sgraband to review as he is most familiar with the original implementation.